Added Float checker for the text fields, So that user can only type 2 places after decimal point. This is for Create UI only. Have to implement the same for Update UI

+++ src/main/java/com/zilogic/dynamicform/Employee.java
@@ -21,8 +21,8 @@ public class Employee implements java.io.Serializable {
     int age;
     @JsonSerializable.JsonElement(name = "DOB", type = GregorianCalendar.class)
     GregorianCalendar dob;
-    @JsonSerializable.JsonElement(name = "EMP ID", type = Integer.class)
-    int id;
+    @JsonSerializable.JsonElement(name = "EMP ID", type = Float.class)
+    float id;
 //
 //    public Employee(String name, int age, GregorianCalendar dob, int id) {
 //        this.name = name;
@@ -43,7 +43,7 @@ public class Employee implements java.io.Serializable {
         return dob;
     }
 
-    public int getId() {
+    public float getId() {
         return id;
     }
